% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_mean.R
\name{avg_mean}
\alias{avg_mean}
\title{Computes the mean of a vector}
\usage{
avg_mean(x, weights = c())
}
\arguments{
\item{x}{A numeric vector.}

\item{weights}{A numeric vector, the same length as x, representing
the weight of each element.}
}
\value{
A scalar value representing the mean of a vector.
}
\description{
Computes the mean of a vector.
}
\section{What it does}{

This function is written to demonstrate a simple principle:
the what we think of as the "average" is a special case of
the dot product. It's just the product of a vector and a
corresponding vector of weights. Here's the formula for
the dot product again:

\deqn{a \cdot b = a_1 b_1 + a_2 b_2 + ... + a_n b_n}
}

\examples{
# For the dot product between two words in a word-document matrix
data(P)
a = P["king",]
avg_mean(a)
}
