% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{correlation}
\alias{correlation}
\alias{most_correlated}
\title{Computes the correlation between two vectors}
\usage{
correlation(x, y)

most_correlated(m, a, num_results = 10)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}
}
\value{
A scalar value representing the correlation between
two vectors.
}
\description{
Computes the correlation between two vectors.
}
\section{What it does}{


\Sexpr[results = rd, stage = build]{
katex::math_to_rd(tex = mathfuncs::correlation())
}

In R code, the above calculation looks like this:\if{html}{\out{<div class="r">}}\preformatted{correlation = function(x, y) \{
   results = covariance(x, y) / sqrt(variance(x) * variance(y))
   return(results)
\}
}\if{html}{\out{</div>}}
}

\examples{
# For the correlation between two terms' distribution in a 
word-document matrix:
data(P)
a = P["king",]
b = P["queen",]
correlation(a, b)

most_correlated(P, a)
}
