% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosine_similarity.R
\name{cosine_similarity}
\alias{cosine_similarity}
\title{Computes the cosine similarity between two vectors}
\usage{
cosine_similarity(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}
}
\value{
A scalar value representing the cosine similarity between
two vectors.
}
\description{
Computes the cosine similarity between two vectors.
}
\section{What it does}{

}

\examples{
# For the cosine between two terms' distribution in a 
word-document matrix:
data(P)
a = P["king",]
b = P["queen",]
cosine_similarity(a, b)

most_similar(P, a)
}
