% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance.R
\name{covariance}
\alias{covariance}
\title{Computes the covariance between two vectors}
\usage{
covariance(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}
}
\value{
A scalar value representing the covariance between
two vectors.
}
\description{
Computes the covariance between two vectors.
}
\section{What it does}{

The covariance is the dot product between the residuals
of two vectors, divided by their length.



\examples{
# For the covariance between two terms' distributions in a 
word-document matrix:
data(P)
a = P["king",]
b = P["queen",]
covariance(a, b)
}
