% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_product.R
\name{dot_product}
\alias{dot_product}
\title{Computes the dot product between two vectors}
\usage{
dot_product(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector, the same length as x.}
}
\value{
A scalar value representing the dot product between two vectors.
}
\description{
Computes the dot product between two vectors.
}



\examples{
# For the dot product between two words in a word-document matrix
data(P)
a = P["king",]
b = P["queen",]
dot_product(a, b)
}
