% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{entropy}
\alias{entropy}
\title{Computes the entropy of a vector}
\usage{
entropy(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A scalar value representing the entropy of a vector.
}
\description{
Computes the entropy of a vector.
}
\examples{
data(P)
a = P["king",]
entropy(a)
}
