% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/least_divergent.R
\name{least_divergent}
\alias{least_divergent}
\title{Computes the relative entropy over a matrix}
\usage{
least_divergent(m, word, num_results = 10)
}
\arguments{
\item{m}{A numeric matrix.}

\item{num_results}{An integer.}

\item{y}{A numeric vector.}
}
\description{
Computes the relative entropy over a matrix
}
\examples{
# For the relative entropy over a word-document matrix:
data(P)
a = P["king",]
least_divergent(P, a)
}
