% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most_correlated.R
\name{most_correlated}
\alias{most_correlated}
\title{Computes the correlation between two vectors}
\usage{
most_correlated(m, a, num_results = 10)
}
\arguments{
\item{num_results}{An integer.}

\item{x}{A numeric vector.}

\item{y}{A numeric vector.}
}
\value{
A scalar value representing the correlation between
two vectors.
}
\description{
Computes the correlation between two vectors.
}
\section{What it does}{


\deqn{\frac{Cov[X, Y]}{\sqrt Var[X] \times Var[Y]}}
}

\examples{
# For the correlation between two terms' distribution in a 
word-document matrix:
data(P)
a = P["king",]
most_correlated(P, a)
}
