% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most_similar.R
\name{most_similar}
\alias{most_similar}
\title{Computes the cosine similarity over a matrix}
\usage{
most_similar(m, a, num_results = 10)
}
\arguments{
\item{m}{A numeric matrix.}

\item{a}{A numeric vector.}

\item{num_results}{An integer.}
}
\description{
Computes the cosine similarity over a matrix.
}
\examples{
# For the cosine similarity over a word-document matrix:
data(P)
a = P["king",]
most_similar(P, a)
}
