% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppmi.R
\name{ppmi}
\alias{ppmi}
\title{Computes the PPMI score for each element of a matrix}
\usage{
ppmi(m)
}
\arguments{
\item{m}{A numeric matrix.}
}
\value{
A matrix representing the PPMI score of each value.
}
\description{
Computes the PPMI score for each element of a matrix.
}
\examples{
data(P)
P_ppmi = ppmi(P)
}
