% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_entropy.R
\name{relative_entropy}
\alias{relative_entropy}
\title{Computes the relative entropy between two vectors}
\usage{
relative_entropy(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector.}
}
\value{
A scalar value representing the relative entropy between
two vectors.
}
\description{
Computes the relative entropy between two vectors.
}
\examples{
# For the relative entropy between two terms' distributions in a 
word-document matrix:
data(P)
a = P["king",]
b = P["queen",]
relative_entropy(a, b)
}
