% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_deviation.R
\name{standard_deviation}
\alias{standard_deviation}
\title{Computes the standard deviation of a vector}
\usage{
standard_deviation(x, mode = "biased")
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A scalar value representing the standard deviation of a vector.
}
\description{
Computes the standard deviation of a vector.
}
\section{What it does}{

The \emph{standard deviation} of a vector is just the square
root of the variance.

\deqn{\sqrt Var[X]}
}

\examples{
# For the standard deviation of a term's distribution in a 
word-document matrix:
data(P)
a = P["king",]
standard_deviation(a)
}
