% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfidf.R
\name{tfidf}
\alias{tfidf}
\title{Computes the TF-IDF score for each element of a matrix}
\usage{
tfidf(m)
}
\arguments{
\item{m}{A numeric matrix.}
}
\value{
A matrix representing the TF-IDF score of each value.
}
\description{
Computes the TF-IDF score for each element of a matrix.
}
\examples{
data(P)
P_tfidf = tfidf(P)
}
