% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{variance}
\alias{variance}
\title{Computes the variance of a vector}
\usage{
variance(x, mode = "biased")
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A scalar value representing the variance of a vector.
}
\description{
Computes the variance of a vector.
}
\section{What it does}{

The variance measures the extent to which the elements
in a sequence of values deviate from their mean. The
"residuals" represent each value \eqn{x_i - \bar x}. For
any given vector, the variance is just the dot product of
the residuals against themselves, divided by \eqn{n} or
\eqn{n+1}.

The \emph{standard deviation} of a vector is just the square
root of the variance.
}

\examples{
# For the variance of a term's distribution in a 
word-document matrix:
data(P)
a = P["king",]
variance(a)
}
