% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore.R
\name{zscore}
\alias{zscore}
\title{Computes the Z-score for each element of a vector}
\usage{
zscore(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector representing the Z-score for each element.
}
\description{
Computes the Z-score for each element of a vector.
}
\section{What it does}{

The \emph{Z}-score of a vector represents how many standard
deviations each value is above or below the mean of that
vector. It is the ratio between a vector's residuals and its
standard deviation.

\deqn{\frac{x_i - \bar x}{\sigma}}
}

\examples{
# For the z-score of a term's distribution in a 
word-document matrix:
data(P)
a = P["king",]
zscore(a)
}
