% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_product.R
\name{dot_product}
\alias{dot_product}
\title{Computes the dot product between two vectors}
\usage{
dot_product(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A numeric vector, the same length as x.}
}
\value{
A scalar value representing the dot product between two vectors.
}
\description{
Computes the dot product between two vectors.
}
\section{What it does}{

The dot product is a mathematical operation that allows you to
compare two vectors of numbers. The dot product is a mathematical
operation at the heart of many descriptive statistics. The dot
is calculated simply by multiply each corresponding element together,
then taking the sum of those elementwise products.

\Sexpr[results = rd, stage = build]{
katex::math_to_rd(tex = mathfuncs::dot_product())
}
}

\examples{
# For the dot product between two words in a word-document matrix
data(P)
a = P["king",]
b = P["queen",]
dot_product(a, b)
}
